Add APIs about OS and OS version here.
