+++ api/server/router/plan/plan.go
@@ -67,5 +67,8 @@ func (t *planRouter) initRoutes(ginEngine *gin.Engine) {
 		planRoute.POST("/:planId/tasks/:taskId", t.runTasks)
 		// 查询任务列表
 		planRoute.POST("/:planId/tasks", t.listTasks)
+
+		// 获取 os 与 os version
+		planRoute.GET("/os", t.getOsList)
 	}
 }
